requirements.txt
feedparser
pyyaml
requests
beautifulsoup4
pandas
matplotlib
